Python/Word_of_the_day_messager/requirements.txt
random
os
sys
logging
twilio.rest
PyDictionary 
pronouncing
word_dataset.txt
